first pass of moving stuff off conftest.py into each test file
