+++ tests/test_aggregates.py
@@ -1,11 +1,22 @@
 """Test for aggregate (group) sensor behavior."""
 
+from collections.abc import AsyncGenerator
 import logging
 from random import randint
 from statistics import mean
+from typing import Any
 
-from homeassistant.components.binary_sensor import DOMAIN as BINARY_SENSOR_DOMAIN
-from homeassistant.components.sensor.const import DOMAIN as SENSOR_DOMAIN
+import pytest
+from pytest_homeassistant_custom_component.common import MockConfigEntry
+
+from homeassistant.components.binary_sensor import (
+    DOMAIN as BINARY_SENSOR_DOMAIN,
+    BinarySensorDeviceClass,
+)
+from homeassistant.components.sensor.const import (
+    DOMAIN as SENSOR_DOMAIN,
+    SensorDeviceClass,
+)
 from homeassistant.const import (
     STATE_OFF,
     STATE_ON,
@@ -14,9 +25,133 @@ from homeassistant.const import (
 )
 from homeassistant.core import HomeAssistant
 
+from .common import get_basic_config_entry_data
+from .conftest import (
+    DEFAULT_MOCK_AREA,
+    init_integration,
+    setup_mock_entities,
+    shutdown_integration,
+)
 from .mocks import MockBinarySensor, MockSensor
+from custom_components.magic_areas.const import (
+    CONF_AGGREGATES_MIN_ENTITIES,
+    CONF_ENABLED_FEATURES,
+    CONF_FEATURE_AGGREGATION,
+    DOMAIN,
+)
 
 _LOGGER = logging.getLogger(__name__)
+
+# Fixtures
+
+
+@pytest.fixture(name="aggregates_config_entry")
+def mock_config_entry_aggregates() -> MockConfigEntry:
+    """Fixture for mock configuration entry."""
+    data = get_basic_config_entry_data(DEFAULT_MOCK_AREA)
+    data.update(
+        {
+            CONF_ENABLED_FEATURES: {
+                CONF_FEATURE_AGGREGATION: {CONF_AGGREGATES_MIN_ENTITIES: 1}
+            }
+        }
+    )
+    return MockConfigEntry(domain=DOMAIN, data=data)
+
+
+@pytest.fixture(name="_setup_integration_aggregates")
+async def setup_integration_aggregates(
+    hass: HomeAssistant,
+    aggregates_config_entry: MockConfigEntry,
+) -> AsyncGenerator[Any]:
+    """Set up integration with secondary states config."""
+
+    await init_integration(hass, [aggregates_config_entry])
+    yield
+    await shutdown_integration(hass, [aggregates_config_entry])
+
+
+# Entities
+
+
+@pytest.fixture(name="entities_binary_sensor_connectivity_multiple")
+async def setup_entities_binary_sensor_connectivity_multiple(
+    hass: HomeAssistant,
+) -> list[MockBinarySensor]:
+    """Create multiple mock sensor and setup the system with it."""
+    nr_entities = 3
+    mock_binary_sensor_entities = []
+    for i in range(nr_entities):
+        mock_binary_sensor_entities.append(
+            MockBinarySensor(
+                name=f"connectivity_sensor_{i}",
+                unique_id=f"connectivity_sensor_{i}",
+                device_class=BinarySensorDeviceClass.CONNECTIVITY,
+            )
+        )
+    await setup_mock_entities(
+        hass, BINARY_SENSOR_DOMAIN, {DEFAULT_MOCK_AREA: mock_binary_sensor_entities}
+    )
+    return mock_binary_sensor_entities
+
+
+@pytest.fixture(name="entities_sensor_temperature_multiple")
+async def setup_entities_sensor_temperature_multiple(
+    hass: HomeAssistant,
+) -> list[MockSensor]:
+    """Create multiple mock sensor and setup the system with it."""
+    nr_entities = 3
+    mock_sensor_entities = []
+    for i in range(nr_entities):
+        random_value = randint(0, 100)
+        mock_sensor_entities.append(
+            MockSensor(
+                name=f"temperature_sensor_{i}",
+                unique_id=f"temperature_sensor_{i}",
+                native_value=random_value,
+                device_class=SensorDeviceClass.TEMPERATURE,
+                native_unit_of_measurement=UnitOfTemperature.CELSIUS,
+                unit_of_measurement=UnitOfTemperature.CELSIUS,
+                extra_state_attributes={
+                    "unit_of_measurement": UnitOfTemperature.CELSIUS,
+                },
+            )
+        )
+    await setup_mock_entities(
+        hass, SENSOR_DOMAIN, {DEFAULT_MOCK_AREA: mock_sensor_entities}
+    )
+    return mock_sensor_entities
+
+
+@pytest.fixture(name="entities_sensor_current_multiple")
+async def setup_entities_sensor_current_multiple(
+    hass: HomeAssistant,
+) -> list[MockSensor]:
+    """Create multiple mock sensor and setup the system with it."""
+    nr_entities = 3
+    mock_sensor_entities = []
+    for i in range(nr_entities):
+        random_value = randint(0, 100)
+        mock_sensor_entities.append(
+            MockSensor(
+                name=f"current_sensor_{i}",
+                unique_id=f"current_sensor_{i}",
+                native_value=random_value,
+                device_class=SensorDeviceClass.CURRENT,
+                native_unit_of_measurement=UnitOfElectricCurrent.AMPERE,
+                unit_of_measurement=UnitOfElectricCurrent.AMPERE,
+                extra_state_attributes={
+                    "unit_of_measurement": UnitOfElectricCurrent.AMPERE,
+                },
+            )
+        )
+    await setup_mock_entities(
+        hass, SENSOR_DOMAIN, {DEFAULT_MOCK_AREA: mock_sensor_entities}
+    )
+    return mock_sensor_entities
+
+
+# Tests
 
 
 async def test_aggregates_binary_sensor_regular(
