+++ tests/test_ble_tracker_monitor.py
@@ -1,13 +1,82 @@
 """Tests for the BLE Tracker feature."""
 
+from collections.abc import AsyncGenerator
+from typing import Any
+
+import pytest
+from pytest_homeassistant_custom_component.common import MockConfigEntry
+
 from homeassistant.components.binary_sensor import DOMAIN as BINARY_SENSOR_DOMAIN
+from homeassistant.components.sensor.const import DOMAIN as SENSOR_DOMAIN
 from homeassistant.const import ATTR_ENTITY_ID, STATE_OFF, STATE_ON, STATE_UNKNOWN
 from homeassistant.core import HomeAssistant
 
-from custom_components.magic_areas.const import ATTR_PRESENCE_SENSORS
+from custom_components.magic_areas.const import (
+    ATTR_PRESENCE_SENSORS,
+    CONF_BLE_TRACKER_ENTITIES,
+    CONF_ENABLED_FEATURES,
+    CONF_FEATURE_BLE_TRACKERS,
+    DOMAIN,
+)
 
+from tests.common import get_basic_config_entry_data
+from tests.conftest import init_integration, setup_mock_entities, shutdown_integration
 from tests.const import DEFAULT_MOCK_AREA
 from tests.mocks import MockSensor
+
+# Fixtures
+
+
+@pytest.fixture(name="ble_tracker_config_entry")
+def mock_config_entry_ble_tracker() -> MockConfigEntry:
+    """Fixture for mock configuration entry."""
+    data = get_basic_config_entry_data(DEFAULT_MOCK_AREA)
+    data.update(
+        {
+            CONF_ENABLED_FEATURES: {
+                CONF_FEATURE_BLE_TRACKERS: {
+                    CONF_BLE_TRACKER_ENTITIES: ["sensor.ble_tracker_1"],
+                }
+            }
+        }
+    )
+    return MockConfigEntry(domain=DOMAIN, data=data)
+
+
+@pytest.fixture(name="_setup_integration_ble_tracker")
+async def setup_integration_ble_tracker(
+    hass: HomeAssistant,
+    ble_tracker_config_entry: MockConfigEntry,
+) -> AsyncGenerator[Any]:
+    """Set up integration with BLE tracker config."""
+
+    await init_integration(hass, [ble_tracker_config_entry])
+    yield
+    await shutdown_integration(hass, [ble_tracker_config_entry])
+
+
+# Entities
+
+
+@pytest.fixture(name="entities_ble_sensor_one")
+async def setup_entities_ble_sensor_one(
+    hass: HomeAssistant,
+) -> list[MockSensor]:
+    """Create one mock sensor and setup the system with it."""
+    mock_ble_sensor_entities = [
+        MockSensor(
+            name="ble_sensor_1",
+            unique_id="unique_ble_sensor",
+            device_class=None,
+        )
+    ]
+    await setup_mock_entities(
+        hass, SENSOR_DOMAIN, {DEFAULT_MOCK_AREA: mock_ble_sensor_entities}
+    )
+    return mock_ble_sensor_entities
+
+
+# Tests
 
 
 async def test_ble_tracker_presence_sensor(
