+++ tests/test_climate.py
@@ -1,6 +1,11 @@
 """Test for climate groups."""
 
+from collections.abc import AsyncGenerator
 import logging
+from typing import Any
+
+import pytest
+from pytest_homeassistant_custom_component.common import MockConfigEntry
 
 from homeassistant.components.binary_sensor import DOMAIN as BINARY_SENSOR_DOMAIN
 from homeassistant.components.climate.const import DOMAIN as CLIMATE_DOMAIN
@@ -8,11 +13,79 @@ from homeassistant.components.switch.const import DOMAIN as SWITCH_DOMAIN
 from homeassistant.const import ATTR_ENTITY_ID, STATE_OFF, STATE_ON
 from homeassistant.core import HomeAssistant
 
+from .common import get_basic_config_entry_data
+from .conftest import (
+    DEFAULT_MOCK_AREA,
+    init_integration,
+    setup_mock_entities,
+    shutdown_integration,
+)
 from .mocks import MockClimate
-
-from tests.const import DEFAULT_MOCK_AREA
+from custom_components.magic_areas.const import (
+    CONF_CLIMATE_GROUPS_TURN_ON_STATE,
+    CONF_ENABLED_FEATURES,
+    CONF_FEATURE_CLIMATE_GROUPS,
+    CONF_FEATURE_PRESENCE_HOLD,
+    CONF_PRESENCE_HOLD_TIMEOUT,
+    DOMAIN,
+    AreaStates,
+)
 
 _LOGGER = logging.getLogger(__name__)
+
+# Fixtures
+
+
+@pytest.fixture(name="climate_groups_config_entry")
+def mock_config_entry_climate_groups() -> MockConfigEntry:
+    """Fixture for mock configuration entry."""
+    data = get_basic_config_entry_data(DEFAULT_MOCK_AREA)
+    data.update(
+        {
+            CONF_ENABLED_FEATURES: {
+                CONF_FEATURE_PRESENCE_HOLD: {CONF_PRESENCE_HOLD_TIMEOUT: 0},
+                CONF_FEATURE_CLIMATE_GROUPS: {
+                    CONF_CLIMATE_GROUPS_TURN_ON_STATE: AreaStates.OCCUPIED.value,
+                },
+            }
+        }
+    )
+    return MockConfigEntry(domain=DOMAIN, data=data)
+
+
+@pytest.fixture(name="_setup_integration_climate_groups")
+async def setup_integration_climate_groups(
+    hass: HomeAssistant,
+    climate_groups_config_entry: MockConfigEntry,
+) -> AsyncGenerator[Any]:
+    """Set up integration with BLE tracker config."""
+
+    await init_integration(hass, [climate_groups_config_entry])
+    yield
+    await shutdown_integration(hass, [climate_groups_config_entry])
+
+
+# Entities
+
+
+@pytest.fixture(name="entities_climate_group_one")
+async def setup_entities_climate_group_one(
+    hass: HomeAssistant,
+) -> list[MockClimate]:
+    """Create one mock climate and setup the system with it."""
+    mock_climate_entities = [
+        MockClimate(
+            name="climate_1",
+            unique_id="unique_climate",
+        )
+    ]
+    await setup_mock_entities(
+        hass, CLIMATE_DOMAIN, {DEFAULT_MOCK_AREA: mock_climate_entities}
+    )
+    return mock_climate_entities
+
+
+# Tests
 
 
 async def test_climate_group_basic(
