+++ tests/test_cover.py
@@ -1,16 +1,84 @@
 """Test for cover groups."""
 
 from collections import defaultdict
+from collections.abc import AsyncGenerator
 import logging
+from typing import Any
+
+import pytest
+from pytest_homeassistant_custom_component.common import MockConfigEntry
 
 from homeassistant.components.cover import CoverDeviceClass
 from homeassistant.components.cover.const import DOMAIN as COVER_DOMAIN
 from homeassistant.const import ATTR_DEVICE_CLASS, ATTR_ENTITY_ID, STATE_OPEN
 from homeassistant.core import HomeAssistant
 
+from .common import get_basic_config_entry_data
+from .conftest import (
+    DEFAULT_MOCK_AREA,
+    init_integration,
+    setup_mock_entities,
+    shutdown_integration,
+)
 from .mocks import MockCover
+from custom_components.magic_areas.const import (
+    CONF_ENABLED_FEATURES,
+    CONF_FEATURE_COVER_GROUPS,
+    DOMAIN,
+)
 
 _LOGGER = logging.getLogger(__name__)
+
+
+# Fixtures
+@pytest.fixture(name="cover_groups_config_entry")
+def mock_config_entry_cover_groups() -> MockConfigEntry:
+    """Fixture for mock configuration entry."""
+    data = get_basic_config_entry_data(DEFAULT_MOCK_AREA)
+    data.update({CONF_ENABLED_FEATURES: {CONF_FEATURE_COVER_GROUPS: {}}})
+    return MockConfigEntry(domain=DOMAIN, data=data)
+
+
+@pytest.fixture(name="_setup_integration_cover_group")
+async def setup_integration_cover_group(
+    hass: HomeAssistant,
+    cover_groups_config_entry: MockConfigEntry,
+) -> AsyncGenerator[Any]:
+    """Set up integration with secondary states config."""
+
+    await init_integration(hass, [cover_groups_config_entry])
+    yield
+    await shutdown_integration(hass, [cover_groups_config_entry])
+
+
+# Entities
+
+
+@pytest.fixture(name="entities_sensor_cover_all_classes_multiple")
+async def setup_entities_sensor_cover_all_classes_multiple(
+    hass: HomeAssistant,
+) -> list[MockCover]:
+    """Create multiple mock sensor and setup the system with it."""
+
+    nr_entities = 3
+    mock_cover_entities = []
+
+    for dc in CoverDeviceClass:
+        for i in range(nr_entities):
+            mock_cover_entities.append(
+                MockCover(
+                    name=f"cover_{dc.value}_{i}",
+                    unique_id=f"cover_{dc.value}_{i}",
+                    device_class=dc.value,
+                )
+            )
+    await setup_mock_entities(
+        hass, COVER_DOMAIN, {DEFAULT_MOCK_AREA: mock_cover_entities}
+    )
+    return mock_cover_entities
+
+
+# Tests
 
 
 async def test_cover_group_basic(
