+++ tests/test_media_player.py
@@ -1,12 +1,17 @@
 """Test for aggregate (group) sensor behavior."""
 
+from collections.abc import AsyncGenerator
 import logging
+from typing import Any
+
+import pytest
+from pytest_homeassistant_custom_component.common import MockConfigEntry
 
 from homeassistant.components.binary_sensor import DOMAIN as BINARY_SENSOR_DOMAIN
-from homeassistant.components.media_player import DOMAIN as MEDIA_PLAYER_DOMAIN
 from homeassistant.components.media_player.const import (
     ATTR_MEDIA_CONTENT_ID,
     ATTR_MEDIA_CONTENT_TYPE,
+    DOMAIN as MEDIA_PLAYER_DOMAIN,
     SERVICE_PLAY_MEDIA,
     MediaType,
 )
@@ -20,17 +25,107 @@ from homeassistant.const import (
 )
 from homeassistant.core import HomeAssistant
 
+from .common import get_basic_config_entry_data
+from .conftest import (
+    DEFAULT_MOCK_AREA,
+    init_integration,
+    setup_mock_entities,
+    shutdown_integration,
+)
 from .mocks import MockBinarySensor, MockMediaPlayer
-from custom_components.magic_areas.const import AreaStates
+from custom_components.magic_areas.const import (
+    CONF_ENABLED_FEATURES,
+    CONF_FEATURE_AREA_AWARE_MEDIA_PLAYER,
+    CONF_NOTIFICATION_DEVICES,
+    CONF_NOTIFY_STATES,
+    DOMAIN,
+    AreaStates,
+)
+
+from tests.const import MockAreaIds
 
 _LOGGER = logging.getLogger(__name__)
+
+# Fixtures
+
+
+@pytest.fixture(name="area_aware_media_player_global_config_entry")
+def mock_config_entry_area_aware_media_player_global() -> MockConfigEntry:
+    """Fixture for mock configuration entry."""
+    data = get_basic_config_entry_data(MockAreaIds.GLOBAL)
+    return MockConfigEntry(domain=DOMAIN, data=data)
+
+
+@pytest.fixture(name="area_aware_media_player_area_config_entry")
+def mock_config_entry_area_aware_media_player_area() -> MockConfigEntry:
+    """Fixture for mock configuration entry."""
+    data = get_basic_config_entry_data(DEFAULT_MOCK_AREA)
+    data.update(
+        {
+            CONF_ENABLED_FEATURES: {
+                CONF_FEATURE_AREA_AWARE_MEDIA_PLAYER: {
+                    CONF_NOTIFICATION_DEVICES: ["media_player.media_player_1"],
+                    CONF_NOTIFY_STATES: [AreaStates.OCCUPIED],
+                }
+            }
+        }
+    )
+    return MockConfigEntry(domain=DOMAIN, data=data)
+
+
+@pytest.fixture(name="_setup_integration_area_aware_media_player")
+async def setup_integration_area_aware_media_player(
+    hass: HomeAssistant,
+    area_aware_media_player_global_config_entry: MockConfigEntry,
+    area_aware_media_player_area_config_entry: MockConfigEntry,
+) -> AsyncGenerator[Any]:
+    """Set up integration with secondary states config."""
+
+    await init_integration(
+        hass,
+        [
+            area_aware_media_player_area_config_entry,
+            area_aware_media_player_global_config_entry,
+        ],
+    )
+    yield
+    await shutdown_integration(
+        hass,
+        [
+            area_aware_media_player_area_config_entry,
+            area_aware_media_player_global_config_entry,
+        ],
+    )
+
+
+# Entities
+
+
+@pytest.fixture(name="entities_media_player_single")
+async def setup_entities_media_player_single(
+    hass: HomeAssistant,
+) -> list[MockMediaPlayer]:
+    """Create multiple mock sensor and setup the system with it."""
+
+    mock_media_player_entities = []
+
+    mock_media_player_entities.append(
+        MockMediaPlayer(name="media_player_1", unique_id="media_player_1")
+    )
+    await setup_mock_entities(
+        hass, MEDIA_PLAYER_DOMAIN, {DEFAULT_MOCK_AREA: mock_media_player_entities}
+    )
+    return mock_media_player_entities
+
+
+# Tests
 
 
 async def test_area_aware_media_player(
     hass: HomeAssistant,
     entities_media_player_single: list[MockMediaPlayer],
     entities_binary_sensor_motion_one: list[MockBinarySensor],
-    _setup_integration_area_aware_media_player,
+    _setup_integration_area_aware_media_player: AsyncGenerator[Any, None],
 ) -> None:
     """Test the area aware media player."""
 
@@ -52,12 +147,14 @@ async def test_area_aware_media_player(
 
     # Area-Aware media player
     area_aware_media_player_state = hass.states.get(area_aware_media_player_id)
+    assert area_aware_media_player_state is not None
     assert area_aware_media_player_state.state == STATE_IDLE
 
     # Test area clear
 
     # Ensure area is clear
     area_state = hass.states.get(area_sensor_entity_id)
+    assert area_state is not None
     assert area_state.state == STATE_OFF
 
     # Send play to AAMP
@@ -77,6 +174,7 @@ async def test_area_aware_media_player(
     assert media_player_state.state == STATE_OFF
 
     area_aware_media_player_state = hass.states.get(area_aware_media_player_id)
+    assert area_aware_media_player_state is not None
     assert area_aware_media_player_state.state == STATE_IDLE
 
     # Test area occupied
@@ -89,7 +187,9 @@ async def test_area_aware_media_player(
     area_binary_sensor = hass.states.get(area_sensor_entity_id)
     motion_sensor = hass.states.get(motion_sensor_entity_id)
 
+    assert motion_sensor is not None
     assert motion_sensor.state == STATE_ON
+    assert area_binary_sensor is not None
     assert area_binary_sensor.state == STATE_ON
     assert AreaStates.OCCUPIED in area_binary_sensor.attributes["states"]
 
@@ -110,6 +210,7 @@ async def test_area_aware_media_player(
     assert media_player_state.state == STATE_PLAYING
 
     area_aware_media_player_state = hass.states.get(area_aware_media_player_id)
+    assert area_aware_media_player_state is not None
     assert area_aware_media_player_state.state == STATE_IDLE
 
     # Turn off area MP
