+++ tests/test_threshold.py
@@ -1,17 +1,105 @@
 """Test for aggregate (group) sensor behavior."""
 
 import asyncio
+from collections.abc import AsyncGenerator
 import logging
+from typing import Any
+
+import pytest
+from pytest_homeassistant_custom_component.common import MockConfigEntry
 
 from homeassistant.components.binary_sensor import DOMAIN as BINARY_SENSOR_DOMAIN
-from homeassistant.components.sensor import DOMAIN as SENSOR_DOMAIN
-from homeassistant.components.threshold.binary_sensor import ATTR_HYSTERESIS, ATTR_UPPER
+from homeassistant.components.sensor.const import (
+    DOMAIN as SENSOR_DOMAIN,
+    SensorDeviceClass,
+)
+from homeassistant.components.threshold.const import ATTR_HYSTERESIS, ATTR_UPPER
 from homeassistant.const import LIGHT_LUX, STATE_OFF, STATE_ON
 from homeassistant.core import HomeAssistant
 
-from .mocks import MockBinarySensor
+from .common import get_basic_config_entry_data
+from .conftest import (
+    DEFAULT_MOCK_AREA,
+    init_integration,
+    setup_mock_entities,
+    shutdown_integration,
+)
+from .mocks import MockBinarySensor, MockSensor
+from custom_components.magic_areas.const import (
+    CONF_AGGREGATES_ILLUMINANCE_THRESHOLD,
+    CONF_AGGREGATES_ILLUMINANCE_THRESHOLD_HYSTERESIS,
+    CONF_AGGREGATES_MIN_ENTITIES,
+    CONF_ENABLED_FEATURES,
+    CONF_FEATURE_AGGREGATION,
+    DOMAIN,
+)
 
 _LOGGER = logging.getLogger(__name__)
+
+# Fixtures
+
+
+@pytest.fixture(name="threshold_config_entry")
+def mock_config_entry_threshold() -> MockConfigEntry:
+    """Fixture for mock configuration entry."""
+    data = get_basic_config_entry_data(DEFAULT_MOCK_AREA)
+    data.update(
+        {
+            CONF_ENABLED_FEATURES: {
+                CONF_FEATURE_AGGREGATION: {
+                    CONF_AGGREGATES_MIN_ENTITIES: 1,
+                    CONF_AGGREGATES_ILLUMINANCE_THRESHOLD: 600,
+                    CONF_AGGREGATES_ILLUMINANCE_THRESHOLD_HYSTERESIS: 10,
+                },
+            }
+        }
+    )
+    return MockConfigEntry(domain=DOMAIN, data=data)
+
+
+@pytest.fixture(name="_setup_integration_threshold")
+async def setup_integration_threshold(
+    hass: HomeAssistant,
+    threshold_config_entry: MockConfigEntry,
+) -> AsyncGenerator[Any]:
+    """Set up integration with secondary states config."""
+
+    await init_integration(hass, [threshold_config_entry])
+    yield
+    await shutdown_integration(hass, [threshold_config_entry])
+
+
+# Entities
+
+
+@pytest.fixture(name="entities_sensor_illuminance_multiple")
+async def setup_entities_sensor_illuminance_multiple(
+    hass: HomeAssistant,
+) -> list[MockSensor]:
+    """Create multiple mock sensor and setup the system with it."""
+    nr_entities = 3
+    mock_sensor_entities = []
+    for i in range(nr_entities):
+        mock_sensor_entities.append(
+            MockSensor(
+                name=f"illuminance_sensor_{i}",
+                unique_id=f"illuminance_sensor_{i}",
+                native_value=0.0,
+                device_class=SensorDeviceClass.ILLUMINANCE,
+                native_unit_of_measurement=LIGHT_LUX,
+                unit_of_measurement=LIGHT_LUX,
+                extra_state_attributes={
+                    "unit_of_measurement": LIGHT_LUX,
+                },
+            )
+        )
+    await setup_mock_entities(
+        hass, SENSOR_DOMAIN, {DEFAULT_MOCK_AREA: mock_sensor_entities}
+    )
+    return mock_sensor_entities
+
+
+# Tests
 
 
 async def test_threshold_sensor_light(
@@ -49,7 +137,7 @@ async def test_threshold_sensor_light(
     for mock_entity in entities_sensor_illuminance_multiple:
         hass.states.async_set(
             mock_entity.entity_id,
-            sensor_threhsold_upper + sensor_hysteresis + 1,
+            str(sensor_threhsold_upper + sensor_hysteresis + 1),
             attributes={"unit_of_measurement": LIGHT_LUX},
         )
     await hass.async_block_till_done()
@@ -66,7 +154,9 @@ async def test_threshold_sensor_light(
     # Reset illuminance sensor values to 0
     for mock_entity in entities_sensor_illuminance_multiple:
         hass.states.async_set(
-            mock_entity.entity_id, 0.0, attributes={"unit_of_measurement": LIGHT_LUX}
+            mock_entity.entity_id,
+            str(0.0),
+            attributes={"unit_of_measurement": LIGHT_LUX},
         )
     await hass.async_block_till_done()
